chore: api 和通知服务文件结构
