+++ tsconfig.json
@@ -7,6 +7,6 @@
   },
   "include": [
     "main.ts",
-    "src/*.ts"
+    "src/**/*.ts"
   ]
 }
